fixed lib->get_libdir for eselect bug workaround Package-Manager: portage-2.1.3.12

+++ sci-libs/lapack-atlas/lapack-atlas-3.7.39.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2007 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/sci-libs/lapack-atlas/lapack-atlas-3.7.39.ebuild,v 1.1 2007/10/10 10:04:59 markusle Exp $
+# $Header: /var/cvsroot/gentoo-x86/sci-libs/lapack-atlas/lapack-atlas-3.7.39.ebuild,v 1.2 2007/10/11 17:04:17 bicatali Exp $
 
 inherit eutils flag-o-matic toolchain-funcs fortran autotools versionator
 
@@ -164,21 +164,12 @@ src_test() {
 
 pkg_postinst() {
 	local current_lib=$(eselect lapack show | cut -d' ' -f2)
-	# uncomment below when eselect bug #189942 is fixed, together with DEPEND
-	#if [[ -z ${current_p} || ${current_p} == ${ESELECT_PROF} ]]; then
-
 	# this snippet works around the eselect bug #189942 and makes
-	# sure that users upgrading from a previous lapack-atlas 
+	# sure that users upgrading from a previous lapack-atlas
 	# version pick up the new pkg-config files
-	if [[ ${current_lib} == ${ESELECT_PROF} ]]; then
-		local configfile="${ROOT}"/etc/env.d/lapack/lib/config
-		if [[ -e ${configfile} ]]; then
-			rm -f ${configfile}
-			eselect lapack set ${ESELECT_PROF}
-		fi
-	fi
-
-	if [[ -z ${current_lib} ]]; then
+	if [[ ${current_lib} == ${ESELECT_PROF} || -z ${current_lib} ]]; then
+		local configfile="${ROOT}"/etc/env.d/lapack/$(get_libdir)/config
+		[[ -e ${configfile} ]] && rm -f ${configfile}
 		eselect lapack set ${ESELECT_PROF}
 		elog "lapack has been eselected to ${ESELECT_PROF}"
 	else
